Layout component updated with new refactors
