+++ src/pages/index.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Layout from '../components/Layout'
+import Layout from '../components/Layout/Layout'
 
 import Header from '../components/Header'
 import Main from '../components/Main'
